requirements.txt
fastapi
uvicorn[standard]
streamlit
python-dotenv
sqlalchemy
aiosqlite
httpx
pydantic
